# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: SpecDriven API
  version: 0.0.1
servers:
  - url: /

security:
  - bearerAuth: []

paths:
  /ping:
    get:
      security: []
      operationId: ping
      summary: Lightweight health check
      responses:
        '200':
          description: Pong
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PingResponse'
        '500':
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    post:
      security:
        - bearerAuth: []
        - {}
      operationId: createUser
      summary: Create a new user
      description: |
        Requires a valid Bearer token **unless** the system is in bootstrap state.

        **Bootstrap exception**: If there are **zero users** in the database, this endpoint may be called without a Bearer token to create the first user.

        After at least one user exists, this endpoint requires `Authorization: Bearer <token>`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    get:
      operationId: listUsers
      summary: List users (paged)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: username
          in: query
          required: false
          schema:
            type: string
          description: Exact match on username.
        - name: emailAddress
          in: query
          required: false
          schema:
            type: string
            format: email
          description: Exact match on email address.
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive partial match on name.
        - name: roleName
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RoleName'
          description: Exact match on role name; returns users that have the role assigned.
      responses:
        '200':
          description: A page of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /users/{userId}:
    get:
      operationId: getUserById
      summary: Get a user by id
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    put:
      operationId: updateUser
      summary: Update a user by id
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: deleteUser
      summary: Delete user by id
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /users/{userId}/roles/{roleName}:
    put:
      operationId: assignRoleToUser
      summary: Assign a role to a user (idempotent)
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RoleNamePath'
      responses:
        '204':
          description: Assigned (idempotent)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      operationId: removeRoleFromUser
      summary: Remove a role from a user (idempotent)
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/RoleNamePath'
      responses:
        '204':
          description: Removed (idempotent)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /login:
    post:
      security: []
      operationId: login
      summary: Login and obtain an authentication token
      description: |
        **Does not require a Bearer token.**

        Gated by feature flag `FeatureFlag.usersApi`.

        For invalid credentials, this operation returns the same 400 error response for unknown username and incorrect password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send `Authorization: Bearer <token>`.

        Token is obtained from `POST /login`.

  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
      description: 1-based page number.
    PageSize:
      name: pageSize
      in: query
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Number of items per page.
    RoleNamePath:
      name: roleName
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/RoleName'

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated caller is not permitted to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Transient failure
      headers:
        Retry-After:
          description: Optional retry delay in seconds.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    PingResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: pong
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, documented error code.
        message:
          type: string
          description: Human-readable summary safe for clients.
        details:
          type: object
          additionalProperties: true
          description: Optional structured details for support/diagnostics.

    RoleName:
      type: string
      description: Known role names.
      enum: [ADMIN, USER, AUDITOR]

    Permission:
      type: string
      description: Known permission values.
      enum: [USER_READ, USER_WRITE, ROLE_ASSIGN]

    Role:
      type: object
      required: [roleName, permissions]
      properties:
        roleName:
          $ref: '#/components/schemas/RoleName'
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
          uniqueItems: true

    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, name, password, emailAddress]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
        name:
          type: string
          minLength: 1
          maxLength: 200
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Password input (stored hashed). Never returned.
        emailAddress:
          type: string
          format: email
          maxLength: 254

    UpdateUserRequest:
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
        name:
          type: string
          minLength: 1
          maxLength: 200
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Password update (stored hashed). Never returned.
        emailAddress:
          type: string
          format: email
          maxLength: 254

    User:
      type: object
      additionalProperties: false
      required: [id, username, name, emailAddress, roles, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        emailAddress:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated

    UserPage:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0

    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 50
          description: Username used to authenticate.
        password:
          type: string
          minLength: 1
          maxLength: 200
          description: Password used to authenticate.

    LoginResponse:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          description: Spring Security token for subsequent requests.
        tokenType:
          type: string
          enum: [Bearer]
          description: Optional token type; when present, always `Bearer`.